# car/api/swagger.yml
openapi: "3.0.3"

info:
  title: Car Service
  version: v1

tags:
  - name: Car
    description: "Everything about your Cars"


paths:
  /cars:
    get:
      summary: Get all cars
      operationId: getAllCars
      tags:
        - Cars
      responses:
        200:
          description: OK - Successfully return a list of Cars
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CarResponse'
    post:
      summary: Create a new Car
      operationId: createCar
      tags:
        - Cars
      requestBody:
        description: A JSON object containing Car information
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarRequest'
      responses:
        200:
          description: Created a new Car
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarResponse'
  /cars/{carId}:
    put:
      summary: Update a Cars by id
      operationId: updateCarsById
      tags:
        - Cars
      parameters:
        - in: path
          name: carId
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: A JSON object containing Car information
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarRequest'
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarResponse'
        404:
          description: Not found
    get:
      summary: Get a Car
      operationId: getCarById
      tags:
        - Cars
      parameters:
        - in: path
          name: carId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarResponse'
        404:
          description: Not found
    delete:
      summary: delete an existing car
      operationId: deleteCarById
      tags:
        - Cars
      parameters:
        - in: path
          name: carId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: No Content
        404:
          description: Not found

components:
  schemas:
    CarRequest:
      type: object
      properties:
        licensePlate:
          type: string
        company:
          type: string
        model:
          type: string

    CarResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        licensePlate:
          type: string
        company:
          type: string
        model:
          type: string